requirements.txt
selenium
webdriver-manager
pandas
pandas-gbq
google-auth